refactor: rename currency type file extension
