+++ src/componentLibrary/CoinPair/CoinPair.tsx
@@ -1,4 +1,5 @@
-import { CoinIcon, CurrencySymbols } from '../';
+import { CurrencySymbols } from 'types/currency';
+import { CoinIcon } from '../';
 import { CoinPairWrapper, CoinPairWrapperProps } from './CoinPair.style';
 
 interface CoinPairProps extends CoinPairWrapperProps {
